requirements.txt
adafruit-blinka
adafruit-circuitpython-rgb-display
matplotlib
Pillow